+++ lib/java/src/main/java/io/github/brenbar/japi/_Types.java
@@ -18,14 +18,14 @@ class UTrait {
 
 class BinaryEncoding {
 
-    public final Map<String, Long> encodeMap;
-    public final Map<Long, String> decodeMap;
+    public final Map<String, Integer> encodeMap;
+    public final Map<Integer, String> decodeMap;
     public final Integer checksum;
 
-    public BinaryEncoding(Map<String, Long> binaryEncoding, Integer checksum) {
+    public BinaryEncoding(Map<String, Integer> binaryEncoding, Integer checksum) {
         this.encodeMap = binaryEncoding;
         this.decodeMap = binaryEncoding.entrySet().stream()
-                .collect(Collectors.toMap(e -> Long.valueOf(e.getValue()), e -> e.getKey()));
+                .collect(Collectors.toMap(e -> e.getValue(), e -> e.getKey()));
         this.checksum = checksum;
     }
 }
